Add topi/rocm/tensor_intrin.py and refactor.
