+++ python/tvm/topi/rocm/dense_mfma.py
@@ -22,6 +22,11 @@ from tvm import te
 import tvm.autotvm as autotvm
 from .. import tag
 from ..util import traverse_inline, get_const_tuple
+from .tensor_intrin import (
+    intrin_mfma_load_matrix,
+    intrin_mfma_store_matrix,
+    intrin_mfma_gemm,
+)
 
 @autotvm.register_topi_compute("dense_mfma.rocm")
 def dense_mfma(cfg, data, weight, bias=None, out_dtype=None):
@@ -211,146 +216,18 @@ def _schedule_dense_mfma(cfg, s, C):
     shared_shedule(AS, AS_align)
     shared_shedule(BS, BS_align)
 
+    def mfma_compute(M, N, K):
+        A = te.placeholder((M, K), name="A", dtype="float16")
+        B = te.placeholder((K, N), name="B", dtype="float16")
+        k = te.reduce_axis((0, K), name="k")
+        return A, B, te.compute(
+            (M, N),
+            lambda i, j: te.sum(A[i, k].astype("float") * B[j, k].astype("float") , axis=[k]),
+            name="C"
+        )
+
     # Lower the inner loop nest down to MFMA hardware intrinsics
     s[AF].tensorize(b_ii, intrin_mfma_load_matrix((mfma_m, mfma_n, mfma_k), "A", strides_src=AS_stride, strides_dst=AF_stride))
-    s[BF].tensorize(o_ii, intrin_mfma_load_matrix((mfma_m, mfma_n, mfma_k), "A", strides_src=BS_stride, strides_dst=BF_stride))
-    s[CF].tensorize(_ii, intrin_mfma_gemm(shape=(mfma_m, mfma_n, mfma_k), dtype="float16", input_scope="local", strides_A=AF_stride, strides_B=BF_stride, strides_C=CF_stride))
+    s[BF].tensorize(o_ii, intrin_mfma_load_matrix((mfma_m, mfma_n, mfma_k), "W", strides_src=BS_stride, strides_dst=BF_stride))
+    s[CF].tensorize(_ii, intrin_mfma_gemm((mfma_m, mfma_n, mfma_k), mfma_compute, input_scope="local", strides_A=AF_stride, strides_B=BF_stride, strides_C=CF_stride))
     s[CS].tensorize(bbi, intrin_mfma_store_matrix(shape=(mfma_m, mfma_n, mfma_k), strides_src=CF_stride, strides_dst=CS_stride))
-
-def intrin_mfma_load_matrix(shape, matrix, thread=None, strides_src=None, strides_dst=None):
-    M, N, K = shape
-    if matrix == "A":
-        row, col = M, K
-    elif matrix == "B":
-        row, col = K, N
-    output_shape = (row, col)
-
-    A = te.placeholder(output_shape, name=matrix, dtype="float16")
-    BA = tvm.tir.decl_buffer(
-        A.shape, A.dtype, scope="shared", offset_factor=1, strides=strides_src
-    )
-
-    C = te.compute(output_shape, lambda i, j: A[i, j], name="C")
-
-    BC = tvm.tir.decl_buffer(
-        C.shape, C.dtype, scope="local", offset_factor=1, strides=strides_dst
-    )
-
-    def intrin_func(ins, outs):
-        ib = tvm.tir.ir_builder.create()
-
-        BA = ins[0]
-        BC = outs[0]
-
-        tx = thread
-        if tx == None:
-            tx = te.thread_axis("threadIdx.x")
-            ib.scope_attr(tx, "thread_extent", 64)
-
-        blk_td = tx % 16
-        offset = tx // 16
-        # TODO(csullivan): Using offset works, but using tx directly does not, fix this
-        if matrix == "A":
-            for blk_id in range(0,4):
-                ib.emit(BC.vstore([0, blk_id], BA.vload([blk_td, blk_id*4 + offset], "float16")))
-        elif matrix == "B":
-            for blk_id in range(0,4):
-                ib.emit(BC.vstore([0, blk_id], BA.vload([blk_id*4 + offset, blk_td], "float16")))
-        return ib.get()
-
-    return te.decl_tensor_intrin(C.op, intrin_func, binds={A: BA, C: BC})
-
-def intrin_mfma_store_matrix(shape, thread=None, strides_src=None, strides_dst=None):
-    M, N, K = shape
-    A = te.placeholder((M, N), name="A", dtype="float32")
-    BA = tvm.tir.decl_buffer(
-        A.shape, A.dtype, scope="local", offset_factor=1, strides=strides_src
-    )
-    C = te.compute((M, N), lambda i, j: A[i, j], name="C")
-    BC = tvm.tir.decl_buffer(
-        C.shape, C.dtype, scope="shared", offset_factor=1, strides=strides_dst
-    )
-
-    def intrin_func(ins, outs):
-        ib = tvm.tir.ir_builder.create()
-        BA = ins[0]
-        BC = outs[0]
-        tx = thread
-        if tx == None:
-            tx = te.thread_axis("threadIdx.x")
-            ib.scope_attr(tx, "thread_extent", 64)
-        vec_width = 4
-        blk_id = tx // 16
-        blk_td = tx % 16
-
-        # TODO(csullivan): Consider TVM change to BufferVar.__getitem__
-        # to convert int to const for quality of life when using vector types.
-        ib.emit(BC.vstore([blk_td, blk_id*vec_width], BA.vload([0, 0], "float32x4")))
-
-        return ib.get()
-
-    return te.decl_tensor_intrin(C.op, intrin_func, binds={A: BA, C: BC})
-
-def intrin_mfma_gemm(shape, dtype, input_scope, strides_A=None, strides_B=None, strides_C=None):
-    M, N, K = shape
-
-    # TODO(csullivan):Replace below with a function to get the correct
-    # llvm function based on shape and type
-    assert M == 16
-    assert N == 16
-    assert K == 16
-    mfma_instr_name = "llvm.amdgcn.mfma.f32.16x16x16f16"
-    llvm_id = tvm.target.codegen.llvm_lookup_intrinsic_id(mfma_instr_name)
-    assert llvm_id != 0
-
-    A = te.placeholder((M, K), name="A", dtype="float16")
-    B = te.placeholder((K, N), name="B", dtype="float16")
-    k = te.reduce_axis((0, K), name="k")
-    C = te.compute(
-        (M, N),
-        lambda i, j: te.sum(A[i, k].astype("float") * B[j, k].astype("float") , axis=[k]),
-        name="C"
-    )
-
-    Ab = tvm.tir.decl_buffer(
-        A.shape, A.dtype, name="Ab", scope=input_scope, offset_factor=1, strides=strides_A
-    )
-    Bb = tvm.tir.decl_buffer(
-        B.shape, B.dtype, name="Bb", scope=input_scope, offset_factor=1, strides=strides_B
-    )
-    Cb = tvm.tir.decl_buffer(
-        C.shape, C.dtype, name="Cb", scope="local", offset_factor=1, strides=strides_C
-    )
-
-    def intrin_func(ins, outs):
-        Ab, Bb = ins
-        (Cb,) = outs
-
-        def init():
-            ib = tvm.tir.ir_builder.create()
-            ib.emit(Cb.vstore([0, 0], tvm.tir.const(0, "float32x4")))
-            return ib.get()
-
-        def update():
-            ib = tvm.tir.ir_builder.create()
-            a_vec = Ab.vload([0, 0], "float16x4")
-            b_vec = Bb.vload([0, 0], "float16x4")
-            c_vec = Cb.vload([0, 0], "float32x4")
-            args_6 = tvm.tir.const(6, "uint32")
-            # Transpose inputs in order to ensure row-major order on the output for
-            # coalesced vector writes.
-            gemm = tvm.tir.call_llvm_pure_intrin(
-                "float32x4",
-                mfma_instr_name,
-                args_6,
-                b_vec,
-                a_vec,
-                c_vec,
-                0,0,0
-            )
-            ib.emit(Cb.vstore([0, 0], gemm))
-            return ib.get()
-
-        return update(), init(), update()
-
-    return te.decl_tensor_intrin(C.op, intrin_func, binds={A: Ab, B: Bb, C: Cb})
